Moved `Rule` structs to separate files
